Handle humidity as a float value (#7)
The Thermometer/Hygrometer reports relativeHumidityMeasurement.humidity
as a float (e.g. 55.0), so the request failed with "value not set or not
an integer". php-smartthings-api-lib now returns humidity as a float
(#1), so align DeviceReading to match.

- DeviceReading + interface: humidity ?int -> ?float (getHumidity too)
- Tests: assert float humidity values and average
- composer.lock: pull the php-smartthings-api-lib humidity fix

+++ src/DeviceReadingInterface.php
@@ -6,7 +6,7 @@ namespace ChristianBrown\GetSmartHomeTemps;
 
 interface DeviceReadingInterface
 {
-    public function getHumidity(): ?int;
+    public function getHumidity(): ?float;
 
     public function getHumidityTimestamp(): ?int;
